fix -薪资发放记录

+++ backend/src/main/java/cc/mrbird/febs/cos/entity/SalaryRecords.java
@@ -97,7 +97,7 @@ public class SalaryRecords implements Serializable {
     /**
      * 住房公积金
      */
-    private Integer housingFund;
+    private BigDecimal housingFund;
 
     /**
      * 实发工资
